third_party/blink/web_tests/platform/generic/accessibility/accessibility-node-reparent-expected.txt

This test makes sure that AccessibilityNodeObjects are properly detached when the node they point to is reparented to a location that allows them to have a renderer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS expectedButtonRole != expectedDetachedRole is true
PASS canvasButtonRole is expectedButtonRole
PASS detachedCanvasButtonRole is expectedDetachedRole
PASS reparentedButtonRole is expectedButtonRole
PASS successfullyParsed is true

TEST COMPLETE
